+++ libexternal/src/external.c
@@ -435,7 +435,7 @@ void RunloopWait(int *r_success)
 }
 
 // IM-2014-07-09: [[ Bug 12225 ]] Add coordinate conversion functions
-void StackToWindowRect(unsigned int p_win_id, MCRectangle32 *x_rect, int *r_success)
+void StackToWindowRect(uintptr_t p_win_id, MCRectangle32 *x_rect, int *r_success)
 {
     char *t_result;
     
@@ -445,12 +445,12 @@ void StackToWindowRect(unsigned int p_win_id, MCRectangle32 *x_rect, int *r_succ
 		return;
 	}
     
-    t_result = (s_operations[OPERATION_STACK_TO_WINDOW_RECT])((const void*)(uintptr_t)p_win_id, x_rect, NULL, r_success);
+    t_result = (s_operations[OPERATION_STACK_TO_WINDOW_RECT])((const void*)p_win_id, x_rect, NULL, r_success);
 	if (t_result != NULL)
 		s_delete(t_result);
 }
 
-void WindowToStackRect(unsigned int p_win_id, MCRectangle32 *x_rect, int *r_success)
+void WindowToStackRect(uintptr_t p_win_id, MCRectangle32 *x_rect, int *r_success)
 {
 	char *t_result;
     
@@ -460,7 +460,7 @@ void WindowToStackRect(unsigned int p_win_id, MCRectangle32 *x_rect, int *r_succ
 		return;
 	}
     
-	t_result = (s_operations[OPERATION_WINDOW_TO_STACK_RECT])((const void*)(uintptr_t)p_win_id, x_rect, NULL, r_success);
+	t_result = (s_operations[OPERATION_WINDOW_TO_STACK_RECT])((const void*)p_win_id, x_rect, NULL, r_success);
 	if (t_result != NULL)
 		s_delete(t_result);
 }
